% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmaxcount.R, R/dmaxcount.all.R, R/pmaxcount.R,
%   R/qmaxcount.R, R/rmaxcount.R
\name{dmaxcount}
\alias{dmaxcount}
\alias{dmaxcount.all}
\alias{pmaxcount}
\alias{qmaxcount}
\alias{rmaxcount}
\title{Probabilty mass function of the maximum-count distribution}
\usage{
dmaxcount(x, size, space, prob = 1, log = FALSE)

dmaxcount.all(max.x, max.size, space, prob = 1, log = FALSE)

pmaxcount(q, size, space, prob = 1, log.p = FALSE, lower.tail = TRUE)

qmaxcount(p, size, space, prob = 1, log.p = FALSE, lower.tail = TRUE)

rmaxcount(n, size, space, prob = 1)
}
\arguments{
\item{space}{The space parameter for the maximum-count distribution (number of bins)}

\item{prob}{The probability parameter for the occupancy distribution (probability of ball occupying its bin)}

\item{log}{A logical value specifying whether results should be returned as log-probabilities}

\item{max.x}{A vector of numeric values to be used as arguments for the probability mass function}

\item{max.size}{The maximum size parameter for the maximum-count distribution (number of balls)}
}
\value{
If all inputs are correctly specified (i.e., parameters are in allowable range) then the output will be a
vector of probabilities/log-probabilities up to the maximum argument values
}
\description{
\code{dmaxcount.all} returns a matrix of probability or log-probability values up to a maximum arguments.
}
\details{
This function computes probabilities or log-probabilities from the probability mass function of the maximum-count
distribution, which is the distribution for the maximum of the counts for the number of balls in a bin in the extended
occupancy problem.  Details of the algorithm in the classical case can be found in the papers below.  The extension
to include the probability parameter is done using the binomial mixture representation of the extended occupancy problem.

Bonetti, M., Corillo, P. and Ogay, A. (2019) Computing the exact distributions of some functions of the ordered multinomial
counts: maximum, minimum, range and sums of order statistics.

Rappeport, M,A. (1968) Algorithms and computational procedures for the application of order statistics to queuing
problems. PhD thesis, New York University.
}
