% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnegocc.R, R/dnegocc.all.R, R/pnegocc.R,
%   R/qnegocc.R, R/rnegocc.R
\name{dnegocc}
\alias{dnegocc}
\alias{dnegocc.all}
\alias{pnegocc}
\alias{qnegocc}
\alias{rnegocc}
\title{Mass function of the negative occupancy distribution}
\usage{
dnegocc(x, space, occupancy, prob = 1, approx = FALSE, log = FALSE)

dnegocc.all(max.x, space, max.occupancy, prob = 1, approx = FALSE, log = FALSE)

pnegocc(
  x,
  space,
  occupancy,
  prob = 1,
  approx = FALSE,
  log.p = FALSE,
  lower.tail = TRUE
)

qnegocc(
  p,
  space,
  occupancy,
  prob = 1,
  approx = FALSE,
  log.p = FALSE,
  lower.tail = TRUE
)

rnegocc(n, space, occupancy, prob = 1)
}
\arguments{
\item{space}{The space parameter for the negative occupancy distribution (number of bins)}

\item{prob}{The probability parameter for the negative occupancy distribution (probability of ball occupying its bin)}

\item{approx}{A logical value specifying whether to use an approximation for the distribution}

\item{log}{A logical value specifying whether results should be returned as log-probabilities}

\item{max.x}{A vector of numeric values to be used as arguments for the mass function}

\item{max.occupancy}{The maximum occupancy parameter for the negative occupancy distribution (number of occupied bins)}
}
\value{
If all inputs are correctly specified (i.e., parameters are in allowable range and arguments are integers)
then the output will be a matrix of probabilities/log-probabilities
}
\description{
\code{dnegocc.all} returns array of probability or log-probability values up to a maximum argument.
}
\details{
This function computes probabilities or log-probabilities from the mass function of the negative occupancy
distribution.  The computation method uses a recursive algorithm from the following paper:

O'Neill, B. (2021) An examination of the negative-occupancy distribution and the coupon-collector distribution.
}
