% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docc.R, R/docc.all.R, R/pocc.R, R/qocc.R,
%   R/rocc.R
\name{docc}
\alias{docc}
\alias{docc.all}
\alias{pocc}
\alias{qocc}
\alias{rocc}
\title{Mass function of the extended occupancy distribution}
\usage{
docc(x, size, space, prob = 1, approx = FALSE, log = FALSE)

\code{docc.all(max.size, space, prob, approx = FALSE, log = FALSE)}

pocc(
  x,
  size,
  space,
  prob = 1,
  approx = FALSE,
  log.p = FALSE,
  lower.tail = TRUE
)

qocc(
  p,
  size,
  space,
  prob = 1,
  approx = FALSE,
  log.p = FALSE,
  lower.tail = TRUE
)

rocc(n, size, space, prob = 1)
}
\arguments{
\item{space}{The space pararmeter for the occupancy distribution (number of bins)}

\item{prob}{The probability parameter for the occupancy distribution (probability of ball occupying its bin)}

\item{approx}{A logical value specifying whether to use the normal approximation to the occupancy distribution}

\item{log}{A logical value specifying whether results should be returned as log-probabilities}

\item{max.size}{The maximum size parameter for the occupancy distribution (number of balls)}
}
\value{
If all inputs are correctly specified (i.e., parameters are in allowable range and arguments are integers)
then the output will be a vector of probabilities/log-probabilities corresponding to the vector argument x
}
\description{
\code{docc.all} returns the probability or log-probability values for the arguments.
}
\details{
This function computes probabilities or log-probabilities from the mass function of the extended occupancy
distribution.  Further details on the distribution can be found in the following paper:

O'Neill, B. (2021) Three distributions in the extended occupancy problem.
}
