% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doccgap.R, R/doccgap.all.R, R/poccgap.R,
%   R/qoccgap.R, R/roccgap.R
\name{doccgap}
\alias{doccgap}
\alias{doccgap.all}
\alias{poccgap}
\alias{qoccgap}
\alias{roccgap}
\title{Mass function of the occupancy-gap distribution}
\usage{
doccgap(
  x,
  size,
  space = NULL,
  occupancy = size,
  prob = NULL,
  scale = NULL,
  log = FALSE
)

doccgap.all(
  size,
  space = NULL,
  max.occupancy = size,
  prob = NULL,
  scale = NULL,
  log = FALSE
)

poccgap(
  x,
  size,
  space = NULL,
  occupancy = size,
  prob = NULL,
  scale = NULL,
  log.p = FALSE,
  lower.tail = TRUE
)

qoccgap(
  p,
  size,
  space = NULL,
  occupancy = size,
  prob = NULL,
  scale = NULL,
  log.p = FALSE,
  lower.tail = TRUE
)

roccgap(n, size, space = NULL, occupancy = size, prob = NULL, scale = NULL)
}
\arguments{
\item{size}{The size parameter for the occupancy-gap distribution (number of balls)}

\item{space}{The space parameter for the occupancy-gap distribution (number of bins)}

\item{prob}{The probability parameter for the occupancy-gap distribution (probability of ball occupying its bin)}

\item{scale}{The scale parameter for the occupancy-gap distribution}

\item{log}{A logical value specifying whether results should be returned as log-probabilities}

\item{max.occupancy}{The maximum occupancy parameter for the occupancy-gap distribution (number of occupied bins)}
}
\value{
If all inputs are correctly specified (i.e., parameters are in allowable range) then the output
will be a matrix of probabilities/log-probabilities
}
\description{
\code{doccgap.all} returns array of probability or log-probability values up to a maximum argument.
}
\details{
This function computes probabilities or log-probabilities from the mass function of the occupancy-gap
distribution.  The computation method uses a recursive algorithm from the following paper:

O'Neill, B. (XXXX) An examination of the occupancy-gap distribution.

Note: The distribution is parameterised by a \code{scale} parameter, but in applied problems in the context
of the extended occupancy problem this parameter is a function of \code{space} and \code{prob} parameters.
The function allows either parameterisation (i.e., the user can either specify the \code{scale} paramater or
both the \code{space} and \code{prob} paramaters).
}
