% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{functions}
\alias{functions}
\title{functions for greta arrays}
\description{
This is a list of functions (mostly from base R) that are
  currently implemented for fitted occupancy models.
}
\details{
\code{predict} generates predictions of occupancy probabilities, detection
    probabilities, and likely detections (sampled as binary detection/nondetection)
    given these probabilities.
}
\section{Usage}{
 \preformatted{

 # predict if newdata are provided as a data.frame
 predict(object, newdata = NULL, type = c("link", "response"), ...)
 
 # predict if newdata are provided as a raster
 spatial_predict(object, newdata = NULL, type = "response", ...)
 
 # extract coefficients from a fitted model
 coef(object, ...)
 
 # summarise a fitted model
 summary(object, ...)
 
 # extract fitted values from a model
 fitted(object, type = c("link", "response"), ...)
 
 # calculate a pseudo-r2 value based on McFadden's r-squared
 r2_calc(object, ...)
 
 # calculate a suite of validation metrics (r2, AUC, DIC)
 calculate_metrics(object, ...)
 
 plot(x, type, names_occ, names_detect, intercept, ...)
 
 plot_pr_occ(object, npred, var_name, label, ...)
 
 plot_pr_detect(object, npred, var_name, label, scale, ...)
 
}
}

\examples{
\dontrun{
 NULL
}
}
