% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.Stirling.R
\name{log.Stirling}
\alias{log.Stirling}
\title{Logarithms of the Stirling numbers of the second kind}
\usage{
\method{log}{Stirling}(n, k, ncp = 0)
}
\arguments{
\item{n}{A vector of non-negative integer values}

\item{k}{A vector of non-negative integer values}

\item{ncp}{Non-centrality parameter (non-negative numeric value)}
}
\value{
If all inputs are correctly specified then the output will be a matrix containing the logarithms of
the Stirling numbers of the second kind
}
\description{
\code{log.Stirling} returns a matrix of the logarithms of the Stirling numbers of the second kind.
}
\details{
This function computes a matrix of the logarithms of the Stirling numbers of the second kind.  The function
allows the user to give a non-centrality parameter for the non-central Stirling numbers.
}
