% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_beds.R
\docType{import}
\name{merge_projections}
\alias{merge_projections}
\title{Merge a list of projections objects}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{projections}{\code{\link[projections]{merge_projections}}}
}}

