% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{methods}
\alias{methods}
\alias{predict.occupancy_model}
\alias{spatial_predict}
\alias{summary.occupancy_model}
\alias{coef.occupancy_model}
\alias{calculate_metrics}
\alias{r2_calc}
\alias{plot.occupancy_model}
\alias{plot_pr_occ}
\alias{plot_pr_detect}
\title{methods for occupancy_models}
\usage{
\method{predict}{occupancy_model}(object, newdata = NULL,
  type = c("link", "response"), ...)

spatial_predict(object, newdata = NULL, type = "response", ...)

\method{summary}{occupancy_model}(object, ...)

\method{coef}{occupancy_model}(object, ...)

calculate_metrics(object, ...)

r2_calc(object, ...)

\method{plot}{occupancy_model}(x, type = "barplot", names_occ = NULL,
  names_detect = NULL, intercept = FALSE, ...)

plot_pr_occ(object, npred = 1000, var_name = NULL, label = NULL, ...)

plot_pr_detect(object, npred = 1000, var_name = NULL, label = NULL,
  scale = NULL, ...)
}
\arguments{
\item{object}{to add}

\item{newdata}{to add}

\item{type}{to add}

\item{\dots}{to add}

\item{x}{to add}

\item{names_occ}{to add}

\item{names_detect}{to add}

\item{intercept}{to add}

\item{npred}{to add}

\item{var_name}{to add}

\item{label}{to add}

\item{scale}{to add}
}
\description{
This is a list of functions (mostly from base R) that are
  currently implemented for fitted occupancy models.
}
\details{
\code{predict} generates predictions of occupancy probabilities, detection
    probabilities, and likely detections (sampled as binary detection/nondetection)
    given these probabilities.
}
\section{Usage}{
 \preformatted{

 # predict if newdata are provided as a data.frame
 predict(object, newdata = NULL, type = c("link", "response"), ...)
 
 # predict if newdata are provided as a raster
 spatial_predict(object, newdata = NULL, type = "response", ...)
 
 # extract coefficients from a fitted model
 coef(object, ...)
 
 # summarise a fitted model
 summary(object, ...)
 
 # extract fitted values from a model
 fitted(object, type = c("link", "response"), ...)
 
 # calculate a pseudo-r2 value based on McFadden's r-squared
 r2_calc(object, ...)
 
 # calculate a suite of validation metrics (r2, AUC, DIC)
 calculate_metrics(object, ...)
 
 # plot the model coefficients
 plot(x, type, names_occ, names_detect, intercept, ...)
 
 # predictive plots of probabilities of occupancy
 plot_pr_occ(object, npred, var_name, label, ...)
 
 # predictive plots of probabilities of detection
 plot_pr_detect(object, npred, var_name, label, scale, ...)
 
}
}

\examples{
\dontrun{

# fit a model to simulated data
mod <- occupancy(response ~ occ_predictor1 + occ_predictor2 + 
                    (1 | occ_random1) + (1 | occ_random2),
                  ~ detect_predictor1 + detect_predictor2 + 
                   (1 | detect_random1),
               site_id = "site",
               survey_id = "survey",
               data = occupancy_data,
               jags_settings = list(n_iter = 1000, n_burnin = 500, n_thin = 2))
               
# plot the model coefficients
par(mfrow = c(2, 1))
plot(mod)

# extract the model coefficients
coef(mod)

# check model fit
calculate_metrics(mod)

# spatial predictions

# plot_pr_occ example

# plot_pr_detect

}
}
