% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSMocc_functions.R
\name{MSMocc_mod}
\alias{MSMocc_mod}
\title{Title}
\usage{
MSMocc_mod(
  occupancy,
  detection,
  data,
  K,
  species.names,
  niter = 1000,
  nchains = 3,
  seed = NULL,
  save_model = FALSE,
  model_name = paste0("MSMocc_model_", Sys.Date()),
  beta_prior = "dnorm(0, 1/2)"
)
}
\arguments{
\item{model_name}{}
}
\value{

}
\description{
Title
}
\examples{

}
