% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_mod.R
\name{occ_mod}
\alias{occ_mod}
\title{Fit single species, single season occupancy model.}
\usage{
occ_mod()
}
\arguments{
\item{data}{}
}
\value{
return
}
\description{
This function fits the vanilla single species, single season
occupancy model first developed by
\href{https://esajournals.onlinelibrary.wiley.com/doi/10.1890/0012-9658\%282002\%29083\%5B2248\%3AESORWD\%5D2.0.CO\%3B2}{MacKenzie
et al. (2002)}.
}
\details{
details
}
\examples{
occ_mod()
}
