% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy.R
\name{occupancy}
\alias{occupancy}
\title{occupancy: Fit and analyze a variety of occupancy models}
\description{
This package provides functions for fitting and analyzing
multiple occupancy models, including false positive occupancy models.
}
\seealso{
Useful links:
\itemize{
\item \href{https://esajournals.onlinelibrary.wiley.com/doi/10.1890/0012-9658\%282002\%29083\%5B2248\%3AESORWD\%5D2.0.CO\%3B2}{MacKenzie
et al. (2002) occupancy model paper}
\item \href{https://esajournals.onlinelibrary.wiley.com/doi/pdf/10.1890/14-1507.1}{Miller
et al. (2015) false positive occupancy model review paper}
}
}
\author{
\strong{Maintainers:}
\itemize{
\item Christian Stratton (\email{christianstratton@montana.edu}) \cr
\item Meridith L. Bartley (\email{mlbartley@usgs.gov}) \cr \cr
}

Authors:
\itemize{
\item Christian Stratton
\item Meridith Bartley
\item Dr. Kathryn Irvine
}
}
