% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pocc.R, R/qocc.R, R/rocc.R
\name{pocc}
\alias{pocc}
\title{Cumulative distribution function of the extended occupancy distribution}
\usage{
\code{pocc(K, size, space, prob, log.p = FALSE, lower.tail = TRUE)}

\code{qocc(p, size, space, prob, log.p = FALSE, lower.tail = TRUE)}

\code{rocc(n, size, space, prob)}
}
\arguments{
\item{K}{A vector of numeric values to be used as arguments for the cumulative distribution function}

\item{size}{The size parameter for the occupancy distribution (number of balls)}

\item{space}{The space pararmeter for the occupancy distribution (number of bins)}

\item{prob}{The probability parameter for the occupancy distribution (probability of ball occupying its bin)}

\item{log.p}{A logical value specifying whether results should be returned as log-probabilities}

\item{lower.tail}{A logical value specifying whether results are from the cumulative distribution function
or the corresponding survival function}

\item{p}{A vector of numeric values to be used as arguments for the quantile function}

\item{size}{The size parameter for the occupancy distribution (number of balls)}

\item{space}{The space pararmeter for the occupancy distribution (number of bins)}

\item{prob}{The probability parameter for the occupancy distribution (probability of ball occupying its bin)}

\item{log.p}{A logical value specifying whether the input arguments are log-probabilities}

\item{lower.tail}{A logical value specifying whether the input arguments are lower-tail probabilities}

\item{n}{The number of observations in the random sample}

\item{size}{The size parameter for the occupancy distribution (number of balls)}

\item{space}{The space pararmeter for the occupancy distribution (number of bins)}

\item{prob}{The probability parameter for the occupancy distribution (probability of ball occupying its bin)}
}
\value{
If all inputs are correctly specified (i.e., parameters are in allowable range and arguments are integers)
then the output will be a vector of probabilities/log-probabilities corresponding to the vector argument K

If all inputs are correctly specified (i.e., parameters are in allowable range and arguments are valid)
then the output will be a vector of quantiles corresponding to the vector argument p

If all inputs are correctly specified (i.e., parameters are in allowable range and arguments are integers)
then the output will be a vector of probabilities/log-probabilities corresponding to the vector argument K
}
\description{
\code{pocc} returns the probability or log-probability values for the arguments.

\code{qocc} returns the probability or log-probability values for the arguments.

\code{rocc} returns random variables from the extended occupancy distribution.
}
\details{
This function computes probabilities or log-probabilities from the cumulative distribution function of the
extended occupancy distribution.  The computation method uses recursive algorithm that operates from a
pre-computed matrix of values stored in a separate object called \code{ROCC_STORED}.  If \code{size} is "large"
(larger than the length of the pre-computed matrix) then the function use the normal approximation to the
occupancy distribution, and gives a warning to this effect.  Details of the algorithm can be found in the
following papers:

O'Neill, B. (2019) The classical occupancy distribution: computation and approximation.
O'Neill, B. (2019) The extended occupancy distribution as a Markov chain.

This function computes quantiles from the quantile function of the extended occupancy distribution.  The
computation method uses recursive algorithm that operates from a pre-computed matrix of values stored in a
separate object called \code{ROCC_STORED}.  If \code{size} is "large" (larger than the length of the
pre-computed matrix) then the function use the normal approximation to the occupancy distribution, and
gives a warning to this effect.  Details of the algorithm can be found in the following papers:

O'Neill, B. (2019) The classical occupancy distribution: computation and approximation.
O'Neill, B. (2019) The extended occupancy distribution as a Markov chain.

This function generates a set of random variables from the extended occupancy distribution using an underlying
random set of pseudo-random numbers.  To avoid underflow problems in the randomisation, the function operates
using the inverse-transformation sampling method in log-probability space.  (This requires the function to
comg-probabilities from the cumulative distribution function of the extended occupancy distribution.  For more
on this aspect of the function, see the details of the \code{pocc} function.)
}
