% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_functions.R
\name{print.occ_mod}
\alias{print.occ_mod}
\title{Print method for \code{occ_mod} class}
\usage{
\method{print}{occ_mod}(x, ...)
}
\arguments{
\item{x}{An object of class \link{occ_mod}}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Print method for \code{occ_mod} class
}
