% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_functions.R
\name{print.summary.occ_mod}
\alias{print.summary.occ_mod}
\title{Print method for \code{summary.occ_mod} class}
\usage{
\method{print}{summary.occ_mod}(x, ...)
}
\arguments{
\item{x}{An object of class \link{summary.occ_mod}}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Print method for \code{summary.occ_mod} class
}
