% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_beds.R
\name{project_beds}
\alias{project_beds}
\title{Project bed occupancy from admissions}
\usage{
project_beds(x, r_los, n_sim = 1)
}
\arguments{
\item{x}{a \code{projections} object storing forecast of daily admissions}

\item{r_los}{function generating random duration of hospitalisation}

\item{n_sim}{The number of times duration of hospitalisation is simulated for
each admission. Defaults to 10. Only relevant for low (<30) numbers of
initial admissions, in which case it helps accounting for the uncertainty
in LoS.}
}
\description{
This projects bed occupancy using admission incidence stored as a
\code{projections} object, and a distribution of length of stay (los). This is a
wrapper around \code{simulate_occupancy}, which essentially applies this function
to different admission trajectories and collects outputs into a single
\code{projections} object.
}
\examples{
if (require(projections))  {

  ## make fake data
  ## each column after the first is a separate forecast
  admissions <- data.frame(
      date = Sys.Date() - 1:10,
      as.data.frame(replicate(30, sample(1:100, 10, replace = TRUE))))

  x <- build_projections(x = admissions[, -1], dates = admissions$date)
  x
  plot(x)

  ## fake LoS; check `distcrete::distcrete` for discretising existing distributions
  r_los <- function(n) rgeom(n, prob = .3)

  ## project bed occupancy
  beds <- project_beds(x, r_los)
  beds
  plot(beds)
}

}
\seealso{
\code{\link{simulate_occupancy}}
}
