% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.ballbin.R
\name{sample.ballbin}
\alias{sample.ballbin}
\title{Generates simulations from the extended balls-in-bins process}
\usage{
\code{sample.ballbin(n, size, space, prob)}
}
\arguments{
\item{n}{The number of simulations of the process}

\item{size}{The size parameter for the occupancy distribution (number of balls)}

\item{space}{The space pararmeter for the occupancy distribution (number of bins)}

\item{prob}{The probability parameter for the occupancy distribution (probability of ball occupying its bin)}

\item{alloc.prob}{(Optional) A probability vector for the allocation probabilities for the bins}
}
\value{
If all inputs are correctly specified (i.e., parameters are in allowable range) then the output will
be a list of class \code{ballbin} containing \code{n} random samples from the process.  If you call \code{summary}
on this object the output will be another list of class \code{summary.ballbin} containing summary statistics
for each random sample from the process.
}
\description{
\code{sample.ballbin} generates simulated data from the extended balls-in-bins process.
}
\details{
This function generates a simulated set of data from the extended balls-in-bins process.  The outcome is an
object of class \code{ballbin} containing the simulations from the process.  The output object contains the
initial bin-allocation and resulting samples from the process.  Calling \code{summary} on the simulation object
creates a new object of class \code{summary.ballbin} containing summary statistics for each sample, including
the bin-counts, effective sample-size, occupancy number, max-count number, and hitting times for each possible
occupancy number.  Each of these objects has a custom printing and plot methods to give user-friendly output.
}
