% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSMocc_functions.R
\name{sim_MSMocc}
\alias{sim_MSMocc}
\title{Simulate data from the multi-species, misclassification count model for a single season.}
\usage{
sim_MSMocc(
  K = 2,
  M = 20,
  M.conf = 2,
  max_j = 10,
  beta_psi = matrix(c(0, 1, 0, 1.2), ncol = K),
  beta_lambda = matrix(c(0, 1, 0, 1.2), ncol = K),
  theta = matrix(c(0.8, 0.2, 0.2, 0.8), byrow = T, ncol = K),
  seed = NULL,
  rand_visits = TRUE
)
}
\arguments{
\item{K}{number of species}

\item{M}{number of sites}

\item{M.conf}{number of sites that have confirmed observations}

\item{max_j}{maximum number of visits to each site. If \code{rand_visits = FALSE}, this value is the number of visits to each site. If
\code{rand_visits = TRUE}, each site is visited a random
(\code{sample(2:max_j, size = 1)}) number of times.}

\item{beta_psi}{vector of regression coefficients used to generate psi}

\item{beta_lambda}{vector of regression coefficients used to generate lambda}

\item{theta}{confusion matrix for automatic call classifications. Rows of theta should sum to one, diagonal elements are correct identification probabilities, first row corresponds to probabilities for calls truly from species 1.}

\item{seed}{optional seed for reproducibility}

\item{rand_visits}{logical; should each site be visited a random number of
times? See details.}
}
\value{
object of class \code{list} containing the following elements: \cr
\itemize{
\item \code{beta_psi} vector of regression coefficients used to generate psi
\item \code{beta_lambda} vector of regression coefficients used to generate lambda
\item \code{psi_cov} matrix of site level covariates
\item \code{lambda_cov} array of detection level covariates; each slice represents a
single covariate
\item \code{psi} vector of derived site level occupancy probabilities
\item \code{lambda} matrix of derived visit level expected count detection parameter
\item \code{z} matrix of latent occupancy states for each site/species
\item \code{Y} array of observed Bernoulli responses for site/visit/species
\item \code{n_visits} vector of number of visits to each site
\item \code{c.conf} how calls confirmed to species K were classified for confirmed visits
\item \code{c.sum.a} and \code{c.sum.b}  unconfirmed calls classified TO species K from sites with some confirmations ("a") or no confirmations ("b")
\item \code{sites.conf} site locations with confirmations
\item \code{sites.unconf.a} and \code{sites.unconf.b} site locations without confirmations ("a" indecates some observations at these sites were confirmed, "b" sites had no confirmations)
\item \code{data} a data frame containing all information necessary to fit the
model
}
}
\description{
This function simulates data from he multi-species, misclassification count occupancy model first developed by
\href{https://besjournals.onlinelibrary.wiley.com/doi/abs/10.1111/2041-210X.13315}{Wright et al. (2020)}.
}
\details{
This function simulates data from the single season, multiple
species occupancy model to estimate occupancy and relative activity for multiple species simultaneously. If \code{rand_visits = TRUE}, each site is visited a random number of times between two and
\code{max_j}. Covariates are drawn from the uniform(0, 1.5) distribution so
that the effect of the direction of each regression coefficient is
intuitive. Note that if covariates are not desired, \code{beta_psi} and
\code{beta_lambda} can be set to intercepts that generate the desired derived
probabilities.
}
\examples{
n.species <- 4
sim <- sim_MSMocc(K = n.species, M = 100, M.conf = 10, max_j = 20,
                  beta_psi = matrix(c(0, 1,
                                      0, 1.1,
                                      0, 1.2,
                                      0, 1.3),
                                    ncol = n.species),
                  beta_lambda = matrix(c(0, 1, 1,
                                         0, 1, 1,
                                         0, 1, 1.2,
                                         0, 1, 1.4),
                                       ncol = n.species),
                  theta = matrix(c(0.7, 0.1, 0.1, 0.1,
                            0.1, 0.7, 0.1, 0.1,
                            0.1, 0.1, 0.7, 0.1,
                            0.1, 0.1, 0.1, 0.7), nrow = n.species),
                  rand_visits = TRUE)
sim$Y
}
