% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_occ.R
\name{sim_occ}
\alias{sim_occ}
\title{Simulate data from the single species, single season occupancy model.}
\usage{
sim_occ(
  M = 20,
  max_j = 10,
  beta_psi = c(0, 1),
  beta_p = c(0, 1),
  seed = NULL,
  rand_visits = TRUE
)
}
\arguments{
\item{M}{number of sites}

\item{max_j}{maximum number of visits to each site. If \code{rand_visits =
 FALSE}, this value is the number of visits to each site. If
\code{rand_visits = TRUE}, each site is visited a random
(\code{sample(2:max_j, size = 1)}) number of times.}

\item{beta_psi}{vector of regression coefficients used to generate psi}

\item{beta_p}{vector of regression coefficients used to generate p}

\item{seed}{optional seed for reproducibility}

\item{rand_visits}{logical; should each site be visited a random number of
times? See details.}
}
\value{
object of class \code{list} containing the following elements: \cr
\itemize{
\item \code{beta_psi} vector of regression coefficients used to generate psi
\item \code{beta_p} vector of regression coefficients used to generate p
\item \code{psi_cov} matrix of site level covariates
\item \code{p_cov} array of detection level covariates; each slice represents a
single covariate
\item \code{psi} vector of derived site level occupancy probabilities
\item \code{p} matrix of derived visit level detection probabilities
\item \code{z} vector of latent occupancy states for each site
\item \code{Y} matrix of observed Bernoulli responses
\item \code{n_visits} vector of number of visits to each site
}
}
\description{
This function simulates data from the single species, single
season occupancy model first developed by
\href{https://esajournals.onlinelibrary.wiley.com/doi/10.1890/0012-9658\%282002\%29083\%5B2248\%3AESORWD\%5D2.0.CO\%3B2}{MacKenzie
et al. (2002)}.
}
\details{
This function simulates data from the vanilla single season, single
species occupancy model using the logit link function. If \code{rand_visits
 = TRUE}, each site is visited a random number of times between two and
\code{max_j}. Covariates are drawn from the uniform(0, 1.5) distribution so
that the effect of the direction of each regression coefficient is
intuitive. Note that if covariates are not desired, \code{beta_psi} and
\code{beta_p} can be set to intercepts that generate the desired derived
probabilities.
}
\examples{
sim <- sim_occ(M = 100, max_j = 20, beta_psi = c(0, 1), beta_p = c(0, 1, 1), rand_visits = TRUE)
sim$Y
}
