% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_occupancy.R
\name{simulate_occupancy}
\alias{simulate_occupancy}
\title{Simulator for projecting bed occupancy}
\usage{
simulate_occupancy(n_admissions, dates, r_los, n_sim = 10)
}
\arguments{
\item{n_admissions}{An \code{integer} vector giving the number of admissions
predicted for each date in \code{dates}.}

\item{dates}{A vector of dates, ideally as \code{Date} but \code{integer} should work too.}

\item{r_los}{A \code{function} with a single parameter \code{n} returning \code{n} \code{integer}
values of lenth of hospital stay (LoS) in days. Ideally, this should come
from a discrete random distribution, such as \code{rexp} or any \code{distcrete}
object.}

\item{n_sim}{The number of times duration of hospitalisation is simulated for
each admission. Defaults to 10. Only relevant for low (<30) numbers of
initial admissions, in which case it helps accounting for the uncertainty
in LoS.}
}
\description{
This function predits bed occupancy from admission data (dates, and numbers
of admissions on these days). Duration of hospitalisation is provided by a
function returning \code{integer} values for the number of days in hospital.
}
\examples{

## make toy duration of hospitalisation (exponential distribution)
if (require(projections)) {
r_duration <- function(n = 1) rexp(n, .2)

## simulate admissions data (daily incidence)
dates <- Sys.Date() - 1:10
admissions <- rpois(10, 100)

beds <- simulate_occupancy(admissions, dates, r_duration, 30)
plot(beds)

}

}
\author{
Thibaut Jombart
}
