% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_mod.R
\name{summary.occ_mod}
\alias{summary.occ_mod}
\title{Summary method for \code{occ_mod} class}
\usage{
\method{summary}{occ_mod}(
  x,
  burnin = nrow(x[["samples"]][[1]])/2,
  waic_type = 2,
  digits = max(3L, getOption("digits") - 3L),
  ...
)
}
\arguments{
\item{x}{An object of class \link{occ_mod}}

\item{burnin}{number of iterations to discard as burnin}

\item{waic_type}{Type of WAIC to calculate (either 1 or 2)}

\item{digits}{number of digits to print in summary}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Summary method for \code{occ_mod} class
}
