% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docc.R
\name{docc}
\alias{docc}
\title{Mass function of the extended occupancy distribution}
\usage{
\code{docc(K, size, space, prob, log.p = FALSE)}
}
\arguments{
\item{K}{A vector of numeric values to be used as arguments for the mass function}

\item{size}{The size parameter for the occupancy distribution (number of balls)}

\item{space}{The space pararmeter for the occupancy distribution (number of bins)}

\item{prob}{The probability parameter for the occupancy distribution (probability of ball occupying its bin)}

\item{log.p}{A logical value specifying whether results should be returned as log-probabilities}
}
\value{
If all inputs are correctly specified (i.e., parameters are in allowable range and arguments are integers)
then the output will be a vector of probabilities/log-probabilities corresponding to the vector argument K
}
\description{
\code{docc} returns the probability or log-probability values for the arguments.
}
\details{
This function computes probabilities or log-probabilities from the mass function of the extended occupancy
distribution.  The computation method uses recursive algorithm that operates from a pre-computed matrix of
values stored in a separate object called \code{ROCC_STORED}.  If \code{size} is "large" (larger than the
length of the pre-computed matrix) then the function use the normal approximation to the occupancy distribution,
and gives a warning to this effect.  Details of the algorithm can be found in the following papers:

O'Neill, B. (2019) The classical occupancy distribution: computation and approximation.
O'Neill, B. (2019) The extended occupancy distribution as a Markov chain.
}
