% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{import}
\name{incidence}
\alias{incidence}
\title{Compute incidence of events from a vector of dates}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{incidence}{\code{\link[incidence]{incidence}}}
}}

