% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_mod.R
\name{occ_mod}
\alias{occ_mod}
\title{Fit the single species, single season site occupancy model.}
\usage{
occ_mod(
  occupancy,
  detection,
  data,
  niter = 1000,
  seed = NULL,
  save_model = FALSE,
  model_name = paste0("model_", Sys.Date())
)
}
\arguments{
\item{occupancy}{model declaration for the occupancy portion of the model
using standard linear model syntax}

\item{detection}{model declaration for the detection portion of the model
using standard linear model syntax}

\item{data}{data from which to create the detection and covariate matrices.
Each row should represent a single visit to a site. See details and examples
for proper formatting}

\item{niter}{number of MCMC iterations}

\item{seed}{optional seed}

\item{save_model}{logical; should a text file containing the model be
exported?}

\item{model_name}{character string defining the name of the text file
describing the model if \code{save_model = TRUE}}
}
\value{
an object of class \code{matrix} containing posterior samples
}
\description{
This function fits the single species, single
season site occupancy model first developed by
\href{https://esajournals.onlinelibrary.wiley.com/doi/10.1890/0012-9658\%282002\%29083\%5B2248\%3AESORWD\%5D2.0.CO\%3B2}{MacKenzie
et al. (2002)}.
}
\details{
This function fits the single season, single species site occupancy
model using the logit link function. The \code{data} should contain columns
named \code{site} and \code{visit}. See examples.
}
\examples{
# simulate data - note structure
sim <- sim_occ(M = 250, max_j = 10, seed = 11201995, rand_visits = FALSE)
data <- sim$data

# fit model
ex <- occ_mod(occupancy = ~ psi_cov1, detection = ~ p_cov1, data = data,
              niter = 10000)
colMeans(ex)
plot(ex[,4], type = "l")
}
