% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy.R
\docType{package}
\name{occupancy}
\alias{occupancy}
\title{occupancy: A package for computing with occupancy distributions.}
\description{
The occupancy package provides standard mass functions, distribution functions, quantile functions
and random generation for various distributions.  These distributions occur in cases where balls are
randomly allocated to bins.
}
