% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_beds.R
\name{project_beds}
\alias{project_beds}
\title{Project bed occupancy from admissions}
\usage{
project_beds(x, r_los, n_sim = 10)
}
\arguments{
\item{x}{a \code{\link[projections:build_projections]{projections}} object
storing one or more integer forecasts of daily admissions.}

\item{r_los}{A \code{function} with a single parameter \code{n} returning \code{n} \code{integer}
values of lenth of hospital stay (LoS) in days. Ideally, this should come
from a discrete random distribution, such as \code{rexp} or any \code{distcrete}
object.}

\item{n_sim}{The number of times duration of hospitalisation is simulated for
each admission. Defaults to 10. Only relevant for low (<30) numbers of
initial admissions, in which case it helps accounting for the uncertainty
in LoS.

@return A \code{\link[projections:build_projections]{projections}} object
that collects the output from the different admission trajectories.}
}
\description{
This function projects bed occupancy using admission incidence and a
distribution of length of stay (los).
}
\examples{
if (require(projections))  {

  ## make fake data
  ## each column after the first is a separate forecast
  admissions <- data.frame(
      date = Sys.Date() - 1:10,
      as.data.frame(replicate(30, sample(1:100, 10, replace = TRUE))))

  x <- build_projections(x = admissions[, -1], dates = admissions$date)
  x
  plot(x)

  ## fake LoS; check `distcrete::distcrete` for discretising existing distributions
  r_los <- function(n) rgeom(n, prob = .3)

  ## project bed occupancy
  beds <- project_beds(x, r_los)
  beds
  plot(beds)
}

}
