% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_beds.R
\name{project_beds}
\alias{project_beds}
\alias{project_beds.projections}
\alias{project_beds.incidence}
\title{Project bed occupancy from admissions}
\usage{
project_beds(x, ...)

\method{project_beds}{projections}(x, r_los, n_sim = 10, last_date = NULL, ...)

\method{project_beds}{incidence}(x, r_los, n_sim = 10, last_date = NULL, ...)
}
\arguments{
\item{x}{Either a \code{\link[projections:build_projections]{projections}}
object storing one or more integer forecasts of admissions or a
\code{\link[incidence:incidence]{incidence}} object of known daily
admissions.}

\item{...}{Additional arguments passed to other methods.}

\item{r_los}{A \code{function} with a single parameter \code{n} returning \code{n} \code{integer}
values of lenth of hospital stay (LoS) in days. Ideally, this should come
from a discrete random distribution, such as \code{rexp} or any \code{distcrete}
object.}

\item{n_sim}{The number of times duration of hospitalisation is simulated for
each admission. Defaults to 10. Only relevant for low (<30) numbers of
initial admissions, in which case it helps accounting for the uncertainty
in LoS.}

\item{last_date}{the last date to simulate until (defaults to the maximum
date of \code{x}).}
}
\value{
A list of \code{\link[projections:build_projections]{projections}}
objects produced from the admission trajectories.  Where \code{x} was a
\code{\link[incidence:incidence]{incidence}} object the list will be of
size one.
}
\description{
This function projects bed occupancy using admission incidence and a
distribution of length of stay (los).
}
\examples{
  ## fake LoS; check \code{\link[distcrete:distcrete]{distcrete::distcrete}}
  ## for discretising existing distributions
  r_los <- function(n) rgeom(n, prob = .3)


  # Incidence input

  ## fake data
  dates <- Sys.Date() - 1:10
  admissions <- sample(1:100, 10, replace = TRUE)
  x <- incidence::incidence(rep(dates, admissions))
  x
  plot(x)

  ## project bed occupancy
  beds <- project_beds(x, r_los)[[1]]
  beds
  plot(beds)


  # Projections input

  ## make fake data - each column after the first is a separate forecast
  admissions <- data.frame(
      date = Sys.Date() - 1:10,
      as.data.frame(replicate(30, sample(1:100, 10, replace = TRUE))))

  x <- build_projections(x = admissions[, -1], dates = admissions$date)
  x
  plot(x)

  ## project bed occupancy
  beds <- project_beds(x, r_los)
  beds <- merge_projections(beds)
  beds
  plot(beds)

}
